Fix NullType width. Move projection into expressions subpackage.
